Split the Python script in rye
