+++ rye/src/cli/add.rs
@@ -14,28 +14,6 @@ use crate::bootstrap::ensure_self_venv;
 use crate::pyproject::{DependencyKind, PyProject};
 use crate::utils::{format_requirement, CommandOutput};
 
-const PACKAGE_FINDER_SCRIPT: &str = r#"
-import sys
-import json
-from unearth.finder import PackageFinder
-from packaging.version import Version
-
-py_ver = sys.argv[1]
-package = sys.argv[2]
-pre = len(sys.argv) > 3 and sys.argv[3] == "--pre"
-
-finder = PackageFinder(
-    index_urls=["https://pypi.org/simple/"],
-)
-if py_ver:
-    finder.target_python.py_ver = tuple(map(int, py_ver.split('.')))
-choices = iter(finder.find_matches(package))
-if not pre:
-    choices = (m for m in choices if not Version(m.version).is_prerelease)
-
-print(json.dumps([x.as_json() for x in choices]))
-"#;
-
 #[derive(Deserialize, Debug)]
 struct Match {
     name: String,
@@ -272,7 +250,7 @@ fn find_best_matches(
     let mut unearth = Command::new(python_path);
     unearth
         .arg("-c")
-        .arg(PACKAGE_FINDER_SCRIPT)
+        .arg(include_str!("../scripts/package_finder.py"))
         .arg(py_ver.unwrap_or(""))
         .arg(&format_requirement(requirement).to_string());
     if pre {
